feat: restructured GuardianLight folders
