+++ support-frontend/assets/pages/[countryGroupId]/guardianLightLanding/components/componentContainer.tsx
